Moving server communication and fetching regions to TripKitAPI
