+++ Sources/TripKitAPI/extensions/Codable+Helpers.swift
