+++ Sources/TripKitAPI/extensions/Optional+OrThrow.swift
