+++ Sources/TripKitAPI/extensions/String+NonEmpty.swift
